+++ frontend/src/Pages/Staff/KitchenInventory/BulkStock.js
@@ -1,21 +1,21 @@
 import React, { useState, useEffect} from 'react';
-import useKitchenStockDisplay from '../../hooks/useKitchenStockDisplay';
-import useDeleteStock from '../../hooks/useDeleteStock';
-import useUpdateStock from '../../hooks/useUpdateStock';
-import KitchenSidebar from '../../components/KitchenSideBar';
+import useBulkStockDisplay from '../../../hooks/Staff/KitchenInventory/useBulkStockDisplay';
+import useDeleteBulkStock from '../../../hooks/Staff/KitchenInventory/useDeleteBulkStock';
+import useUpdateBulkStock from '../../../hooks/Staff/KitchenInventory/useUpdateBulkStock';
+import KitchenSidebar from '../../../components/KitchenSideBar';
 
-function KitchenInventory () {
-    const {StockList, isLoading, error} = useKitchenStockDisplay();
-    const { deleteStock } = useDeleteStock();
-    const { updateStock } = useUpdateStock();
+function BulkStock () {
+    const {BStockList, isLoading, error} = useBulkStockDisplay();
+    const { deleteBulkStock } = useDeleteBulkStock();
+    const { updateBulkStock } = useUpdateBulkStock();
     const [nameToDelete, setNameToDelete] = useState("");
     const [nameToUpdate, setNameToUpdate] = useState("");
-    const [saveUpdate, setSaveUpdate] = useState("");
-    const [name, setName] = useState("");
-    const [category, setCategory] = useState("");
-    const [quantity, setQuantity] = useState("");
-    const [price, setPrice] = useState("");
-    const [description, setDescription] = useState("");
+    const [bname, setBName] = useState("");
+    const [bcategory, setBCategory] = useState("");
+    const [bquantity, setBQuantity] = useState("");
+    const [bunits, setBUnits] = useState("");
+    const [bprice, setBPrice] = useState("");
+    const [bdescription, setBDescription] = useState("");
     const [searchkey,setsearchkey]=useState('');
     const [filterCategory, setFilterCategory] = useState('');
     const [sort, setSort] = useState('');
@@ -23,11 +23,11 @@ function KitchenInventory () {
     //display only unique categories in filter
     useEffect(() => {
       // Fetch unique categories when StockList changes
-      const uniqueCategories = [...new Set(StockList.map(Stock => Stock.category))];
+      const uniqueCategories = [...new Set(BStockList.map(BulkStock => BulkStock.category))];
       if (uniqueCategories.length > 0) {
         setFilterCategory(''); // Select the first category by default
       }
-    }, [StockList]);
+    }, [BStockList]);
 
 
     if (isLoading) {
@@ -38,31 +38,32 @@ function KitchenInventory () {
         return <div>Error: {error}</div>;
       }
       const handleDelete = async () => {
-        await deleteStock(nameToDelete);
+        await deleteBulkStock(nameToDelete);
        
         setNameToDelete("");
       };
 
-      const getUpdateStock = (Stock) => {
-        setNameToUpdate(Stock._id);
-        setName(Stock.name);
-        setCategory(Stock.category);
-        setQuantity(Stock.quantity);
-        setPrice(Stock.price);
-        setDescription(Stock.description);
+      const getUpdateBulkStock = (BulkStock) => {
+        setNameToUpdate(BulkStock._id);
+        setBName(BulkStock.bname);
+        setBCategory(BulkStock.bcategory);
+        setBQuantity(BulkStock.bquantity);
+        setBUnits(BulkStock.bunits);
+        setBPrice(BulkStock.bprice);
+        setBDescription(BulkStock.bdescription);
       };
 
       const updateDetails = async () => {
-        await updateStock(nameToUpdate,name, category,quantity, price,description );
+        await updateBulkStock(nameToUpdate,bname, bcategory,bquantity,bunits, bprice,bdescription );
       };
 
   
 
       //search and filter
-      const filteredStockList = StockList.filter(Stock => {
+      const filteredStockList = BStockList.filter(BulkStock => {
         return(
-          Stock.name.toLowerCase().includes(searchkey.toLowerCase()) &&
-          (!filterCategory || Stock.category.toLowerCase() === filterCategory.toLowerCase())
+          BulkStock.bname.toLowerCase().includes(searchkey.toLowerCase()) &&
+          (!filterCategory || BulkStock.bcategory.toLowerCase() === filterCategory.toLowerCase())
         );
       }
       );
@@ -70,13 +71,13 @@ function KitchenInventory () {
       const sortData = () => {
         const sortedList = [...filteredStockList];
         if (sort === 'priceAsc') {
-          sortedList.sort((a, b) => a.price - b.price);
+          sortedList.sort((a, b) => a.bprice - b.bprice);
         } else if (sort === 'priceDesc') {
-          sortedList.sort((a, b) => b.price - a.price);
+          sortedList.sort((a, b) => b.bprice - a.bprice);
         } else if (sort === 'quantityAsc') {
-          sortedList.sort((a, b) => a.quantity - b.quantity);
+          sortedList.sort((a, b) => a.bquantity - b.bquantity);
         } else if (sort === 'quantityDesc') {
-          sortedList.sort((a, b) => b.quantity - a.quantity);
+          sortedList.sort((a, b) => b.bquantity - a.bquantity);
         }
         else if (sort === 'newestCreated') {
           sortedList.sort((a, b) => new Date(b.createdAt) - new Date(a.createdAt));
@@ -94,12 +95,12 @@ function KitchenInventory () {
     
 return (
   <div className="row p-0">
-            <KitchenSidebar/>
-            <div className="col">
+        <KitchenSidebar/>
+        <div className="col">
     <div>
-      <h1 className="mb-4 mt-5">Kitchen Inventory</h1>
-      <a href="/AddStock" className="btn btn-primary mb-5">
-              Add New Stock
+      <h1 className="mb-4 mt-5">Bulk Inventory</h1>
+      <a href="/AddBulkStock" className="btn btn-primary mb-5">
+              Add New Bulk Stock
             </a>
 
        {/* Search Input  and filter*/}
@@ -120,8 +121,8 @@ return (
               onChange={(e) => setFilterCategory(e.target.value)}
             >
               <option value="">All Categories</option>
-              {StockList.map(Stock => Stock.category).filter((value, index, self) => self.indexOf(value) === index).map(category => (
-                <option key={category} value={category}>{category}</option>
+              {BStockList.map(BulkStock => BulkStock.bcategory).filter((value, index, self) => self.indexOf(value) === index).map(bcategory => (
+                <option key={bcategory} value={bcategory}>{bcategory}</option>
               ))}
             </select>
           </div>
@@ -168,13 +169,16 @@ return (
               Product Category
             </th>
             <th className="border border-black" scope="col">
-              Quantity Available
+              Quantity 
             </th>
             <th className="border border-black" scope="col">
-              Latest Purchased Price At
+              No of units available
             </th>
             <th className="border border-black" scope="col">
-              SpeciaNotes
+              Latest Purchased Price At per unit
+            </th>
+            <th className="border border-black" scope="col">
+              Special Notes
             </th>
             <th className="border border-black" scope="col">
               Added Date and Time
@@ -187,95 +191,111 @@ return (
           </tr>
 
           
-          {sortData().map((Stock) => (
-            <tbody key={Stock._id}>
+          {sortData().map((BulkStock) => (
+            <tbody key={BulkStock._id}>
               <tr>
                 <td>
-                  {nameToUpdate === Stock._id ? (
+                  {nameToUpdate === BulkStock._id ? (
                     <input
                       class="tabledit-input form-control input-sm"
                       type="text"
-                      name="name"
-                      defaultValue={Stock.name}
+                      name="bname"
+                      defaultValue={BulkStock.bname}
                       disabled=""
                       onChange={(e) => {
-                        setName(e.target.value);
+                        setBName(e.target.value);
                       }}
                     ></input>
                   ) : (
-                    <td>{Stock.name}</td>
+                    <td>{BulkStock.bname}</td>
                   )}
                 </td>
 
                 <td>
-                  {nameToUpdate === Stock._id ? (
+                  {nameToUpdate === BulkStock._id ? (
                     <input
                       class="tabledit-input form-control input-sm"
                       type="text"
-                      name="category"
-                      defaultValue={Stock.category}
+                      name="bcategory"
+                      defaultValue={BulkStock.bcategory}
                       disabled=""
                       onChange={(e) => {
-                        setCategory(e.target.value);
+                        setBCategory(e.target.value);
                       }}
                     ></input>
                   ) : (
-                    <td>{Stock.category}</td>
+                    <td>{BulkStock.bcategory}</td>
                   )}
                 </td>
 
                 <td>
-                  {nameToUpdate === Stock._id ? (
+                  {nameToUpdate === BulkStock._id ? (
                     <input
                       class="tabledit-input form-control input-sm"
                       type="number"
-                      name="quantity"
-                      defaultValue={Stock.quantity}
+                      name="bquantity"
+                      defaultValue={BulkStock.bquantity}
                       disabled=""
                       onChange={(e) => {
-                        setQuantity(e.target.value);
+                        setBQuantity(e.target.value);
                       }}
                     ></input>
                   ) : (
-                    <td>{Stock.quantity}</td>
+                    <td>{BulkStock.bquantity}</td>
                   )}
                 </td>
-
                 <td>
-                  {nameToUpdate === Stock._id ? (
+                  {nameToUpdate === BulkStock._id ? (
                     <input
                       class="tabledit-input form-control input-sm"
                       type="number"
-                      name="price"
-                      defaultValue={Stock.price}
+                      name="bunits"
+                      defaultValue={BulkStock.bunits}
                       disabled=""
                       onChange={(e) => {
-                        setPrice(e.target.value);
+                        setBQuantity(e.target.value);
                       }}
                     ></input>
                   ) : (
-                    <td>{Stock.price}</td>
+                    <td>{BulkStock.bunits}</td>
                   )}
                 </td>
 
                 <td>
-                  {nameToUpdate === Stock._id ? (
+                  {nameToUpdate === BulkStock._id ? (
+                    <input
+                      class="tabledit-input form-control input-sm"
+                      type="number"
+                      name="bprice"
+                      defaultValue={BulkStock.bprice}
+                      disabled=""
+                      onChange={(e) => {
+                        setBPrice(e.target.value);
+                      }}
+                    ></input>
+                  ) : (
+                    <td>{BulkStock.bprice}</td>
+                  )}
+                </td>
+
+                <td>
+                  {nameToUpdate === BulkStock._id ? (
                     <input
                       class="tabledit-input form-control input-sm"
                       type="text"
                       name="description"
-                      defaultValue={Stock.description}
+                      defaultValue={BulkStock.bdescription}
                       disabled=""
                       onChange={(e) => {
-                        setDescription(e.target.value);
+                        setBDescription(e.target.value);
                       }}
                     ></input>
                   ) : (
-                    <td>{Stock.description}</td>
+                    <td>{BulkStock.bdescription}</td>
                   )}
                 </td>
-                <td>{new Date(Stock.createdAt).toLocaleString()}</td>
-                <td>{new Date(Stock.updatedAt).toLocaleString()}</td>
+                <td>{new Date(BulkStock.createdAt).toLocaleString()}</td>
+                <td>{new Date(BulkStock.updatedAt).toLocaleString()}</td>
 
                 <td>
                   <a
@@ -283,14 +303,14 @@ return (
                     className="btn btn-danger"
                     data-bs-toggle="modal"
                     data-bs-target="#Modal"
-                    onClick={() => setNameToDelete(Stock.name)}
+                    onClick={() => setNameToDelete(BulkStock.bname)}
                   >
                     DELETE
                   </a>
                 </td>
 
                 <td>
-                  {nameToUpdate === Stock._id ? (
+                  {nameToUpdate === BulkStock._id ? (
                     <a
                       href="#"
                       className="btn btn-primary"
@@ -302,7 +322,7 @@ return (
                     <a
                       href="#"
                       className="btn btn-primary"
-                      onClick={() => getUpdateStock(Stock)}
+                      onClick={() => getUpdateBulkStock(BulkStock)}
                     >
                       Update
                     </a>
@@ -366,6 +386,4 @@ return (
 
 }
 
-export default KitchenInventory;
-
- 
+export default BulkStock;
