+++ frontend/src/hooks/Staff/KitchenInventory/useDeleteBulkStock.js
