+++ frontend/src/hooks/Staff/KitchenInventory/useDeleteStock.js
